feat(kp-cloud): 业务修改

+++ kp-cloud/kp-system/src/main/java/org/dromara/kp/system/domain/KpConnector.java
@@ -45,7 +45,7 @@ public class KpConnector extends TenantEntity {
     /**
      * 枪号
      */
-    private Long connectorNo;
+    private Integer connectorNo;
 
     /**
      * 枪名称
