+++ kp-cloud/kp-system/src/main/java/org/dromara/kp/system/domain/KpEquipment.java
@@ -120,7 +120,7 @@ public class KpEquipment extends TenantEntity {
     /**
      * 联网类型
      */
-    private Long netType;
+    private Integer netType;
 
     /**
      * 最近上线时间
